requirements.txt
pymongo~=4.7.3
google.generativeai
reportlab~=3.6.8
pypdf~=4.2.0
pandas~=2.2.2
